requirements.txt
streamlit>=1.28.0
assemblyai>=0.5.0
openai>=1.0.0
python-dotenv>=1.0.0
pandas>=1.5.0
uuid>=0.1.0
python-jose>=3.3.0